support "$request_uri" as a log variable
This was documented in the README but never implemented.  Some
popular web servers set REQUEST_URI even though it's not
required by Rack, so allow this variable to be used if possible.

As a side effect, it is also less likely to be modified by
certain handlers (*cough*Rails::Rack::Static*cough*).