Ensure underlying logger objects are sync=true
Userspace buffering defaults are dangerous as the Ruby default
IO objects do not do line-aware buffering.  This makes the
README examples with File.open much safer to use out-of-the-box
for users of the pure-Ruby version.  For users on the MRI C
extension logging to regular files, this should not have any
effect as we've optimized those to do unbuffered write(2)
syscalls anyways.