clogger 0.3.0
You may now force the :reentrant flag to +true+ or +false+ in
your Rack configuration file:

   use Clogger,
       :format => Clogger::Format::Combined,
       :logger => ::File.open("/path/to/log", "ab"),
       :reentrant => true

This is to be compatible with Rack servers that do not use use
events or Fibers for concurrency instead of threads.  By
default, reentrancy is enabled if env["rack.multithread"] is
true, but "rack.multithread" cannot indicative of a reentrancy
requirement in a web server.