when wrapping the body, body.close writes the log
We no longer write the log out at the end of the body.each call.
This is a behavioral change, but fortunately all Rack servers
I've seen call body.close inside an ensure.

This allows us to later pass along the "to_path" method
and not rely on "each" to write the log.