clogger 0.6.0 - :path shortcut, minor bugfixes
Clogger may now be initialized with a :path instead of :logger
to avoid typing "::File" in config.ru files to avoid conflicting
with the completely-unrelated Rack::File.

$request_time is now uses the monotonic clock, making it immune
to system clock changes.  $usec and $msec statements were
completely broken in the pure-Ruby code and now fixed.

Rubinius is supported by the optional C extension as well,
the pure-Ruby code always worked.