test case to simulate Rack::BodyProxy usage pattern
The use of Rack::BodyProxy#method_missing causes failures
under mainline Ruby 1.9.2 and 1.9.1, but not 1.9.3.   This
test case exists to document this (now-fixed) bug for users
stuck on older Rubies.

On a side note, our usage of rb_iterate() should be rewritten
to use rb_block_call() as rb_iterate() is deprecated in 1.9.

A Ruby 1.9.2-p290 user privately reported this issue.