clogger 1.4.0 - updated website URL, cleanups
This updates the documentation to point to the new homepage
at http://clogger.bogomips.org/

This also fixes a minor incompatibility which prevented the GVL from
being released on ruby-trunk (2.2.0dev) during disk operations.

The mailing list continues to be hosted at librelist, but that will
probably soon change to a public-inbox + mailing list.

There are also minor code cleanups.