doc use HTTPS for URLs and move homepage
Thanks to Let's Encrypt, the bogomips.org now supports
HTTPS so our homepage can be moved to:

	https://bogomips.org/clogger/

While clogger.bogomips.org is part of the TLS certificate, it is
needless subjectAltName bloat that should not have been created.
This will also make it easier to have multiple homepages in
the future as a Tor hidden service.