ext: reduce frozen string marking overhead
Using rb_global_variable excessively can be expensive since it
uses a singly-linked list to track addresses.  Since these
strings are all frozen and constant, put them into an array
instead and only mark the array as a global to improve locality.

Ruby 1.9+ has rb_gc_register_mark_object but it is
not part of the documented, public API, so we're not
using it, yet.